refactor(Car.java): Referee 클래스를 생성해서 우승자를 판단하는 기능 분리
